Move UsageFactory away from Rules
